:truck: Refactor: 스터디그룹 관련 exceptions 폴더 별도로 생성 및 구분
