+++ BE_SIDE/study/src/main/java/sw/study/exception/studyGroup/MaxStudyGroupException.java
@@ -1,4 +1,4 @@
-package sw.study.exception;
+package sw.study.exception.studyGroup;
 
 public class MaxStudyGroupException extends RuntimeException {
     public MaxStudyGroupException(String message) {super(message);}
