+++ BE_SIDE/study/src/main/java/sw/study/exception/studyGroup/PermissionDeniedException.java
@@ -1,4 +1,4 @@
-package sw.study.exception;
+package sw.study.exception.studyGroup;
 
 public class PermissionDeniedException extends RuntimeException{
     public PermissionDeniedException(String message) {super(message);}
