+++ BE_SIDE/study/src/main/java/sw/study/exception/studyGroup/StudyGroupFullException.java
@@ -1,4 +1,4 @@
-package sw.study.exception;
+package sw.study.exception.studyGroup;
 
 public class StudyGroupFullException extends RuntimeException{
     public StudyGroupFullException(String message) {super(message);}
