+++ BE_SIDE/study/src/main/java/sw/study/exception/studyGroup/StudyGroupNotFoundException.java
@@ -1,4 +1,4 @@
-package sw.study.exception;
+package sw.study.exception.studyGroup;
 
 public class StudyGroupNotFoundException extends RuntimeException{
     public StudyGroupNotFoundException(String message) {super(message);}
