+++ BE_SIDE/study/src/main/java/sw/study/exception/studyGroup/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package sw.study.exception;
+package sw.study.exception.studyGroup;
 
 public class UnauthorizedException extends RuntimeException{
     public UnauthorizedException(String message){super(message);}
